Reorganizes RPC code
To separate out the RPC core and the RPC Implementation
