+++ Core/Inc/reporting_functions.h
@@ -14,6 +14,7 @@ namespace \
 #define NUM_EVENT_TYPES_KEYPAD 4 // Key_Press, Short_Press_Release, LongPress_Release and Movement
 #define NUM_APPLICATIONS_KEYPAD 16 // Can store between 16 application specific reporting options
 
+
 namespace ReportingEventTypes{
 	enum ReportingEventTypes:uint8_t{
 		KEY_DOWN = 0,
@@ -86,112 +87,4 @@ struct Keypad_Event_Table_Entry_Typedef {
 	uint8_t parameters[];
 };
 
-#define CFG_ENTRY_SERIALIZATION_SIZE_LIMIT 60
 
-template <ReportingFunctionEnum report_type>
-std::array<uint8_t, CFG_ENTRY_SERIALIZATION_SIZE_LIMIT> C(Reporting_Func_Params_Typedef<report_type> obj){
-	std::array<uint8_t, CFG_ENTRY_SERIALIZATION_SIZE_LIMIT> a;
-	a[0] = (uint8_t) report_type;
-	uint8_t siz = (uint8_t) sizeof(Reporting_Func_Params_Typedef<report_type>);
-	a[1] = siz;
-	uint8_t* ptr = reinterpret_cast<uint8_t*>(&obj);
-	for (uint8_t i = 0; i < siz; i++) {
-		a[i + 2] = ptr[i];
-	}
-	return a;
-}
-
-
-struct User_Config_TypeDef{
-	uint8_t application_id;
-	uint8_t keypad_key;
-	std::array<uint8_t, CFG_ENTRY_SERIALIZATION_SIZE_LIMIT> key_down;
-	std::array<uint8_t, CFG_ENTRY_SERIALIZATION_SIZE_LIMIT> mouse_movement;
-	std::array<uint8_t, CFG_ENTRY_SERIALIZATION_SIZE_LIMIT> short_release;
-	std::array<uint8_t, CFG_ENTRY_SERIALIZATION_SIZE_LIMIT> long_release;
-};
-
-
-#define CFG(X, ...) C(Reporting_Func_Params_Typedef<ReportingFunctionEnum::X>__VA_ARGS__)
-
-/*
-User_Config_TypeDef list_of_config_key_values[] =
-{
-		{
-				.application_id = 0,
-				.keypad_key = 0,
-				.key_down = CFG(NO_REPORT,{}),
-				.mouse_movement = CFG(ALTERED_MOUSE_MOVEMENT, 
-					{.x_factor = 1,
-					.y_factor = 1,
-					.z_factor = 1, }),
-				.short_release = CFG(NO_REPORT,{}),
-				.long_release = CFG(NO_REPORT,{}),
-		},
-};
-*/
-
-// Read HID Report from the computer
-// cmd = byte 1
-// if cmd == RPC:
-// 		do_RPC(data)
-
-// do_RPC(){
-// lookup RPC function
-// index = Data[1,2]
-// RPC_Func = RPC_index[index]
-// RPC_Func(data[3:])
-//
-
-//write_to_flash(key,length,value);
-
-
-
-
-
-
-	/* Some other previously tried configuration DSL prototypes
-	 * */
-	/*
-	using no_report = Reporting_Func_Params_Typedef<ReportingFunctionEnum::NO_REPORT>;
-	using altered_mouse_movement = Reporting_Func_Params_Typedef<ReportingFunctionEnum::ALTERED_MOUSE_MOVEMENT>;
-
-	{
-						0,
-						0,
-						C(no_report{}),
-						C(altered_mouse_movement{
-							.x_factor = 1,
-							.y_factor = 1,
-							.z_factor = 1, }),
-						C(no_report{}),
-						C(no_report{}),
-				},
-
-							{
-							.application_id = 0,
-							.keypad_key = 0,
-							.key_down = C(no_report{}),
-							.mouse_movement = C(
-								 altered_mouse_movement{
-									.x_factor = 1,
-									.y_factor = 1,
-									.z_factor = 1, }),
-							.short_release = C(no_report{}),
-							.long_release = C(no_report{}),
-				},
-
-				{
-						0,
-						0,
-						CFG(NO_REPORT,{}),
-						CFG(ALTERED_MOUSE_MOVEMENT, {
-							.x_factor = 1,
-							.y_factor = 1,
-							.z_factor = 1, }),
-						CFG(NO_REPORT,{}),
-						CFG(NO_REPORT,{}),
-				},
-	}
-
-	*/
