+++ Core/Src/main.cpp
@@ -27,7 +27,7 @@
 #include "mouse_event_handler.hpp"
 #include "circular_buffers.hpp"
 #include "key_value_tree.hpp"
-#include "rpc.hpp"
+#include "rpc_impl.hpp"
 
 
 /* Private includes ----------------------------------------------------------*/
@@ -178,7 +178,7 @@ Flash_Key_Value_Tree r_tree = Flash_Key_Value_Tree((uint32_t)flash_config_tree_r
 
 MouseEventHandler mouse_event_handler(&stop_keypress_timer, &start_keypress_timer, &read_keypress_time_ms);
 
-RPC hid_rpc_obj = RPC(&r_tree);
+auto hid_rpc_obj = RPC<RPC_Function_Enum, num_rpc_impl_funcs>(&r_tree);
 
 
 void spi_rx_complete(SPI_HandleTypeDef *hspi){
